+++ gazebo/gui/CloneWindow.hh
@@ -30,7 +30,8 @@ namespace gazebo
   {
     class CloneWindowPrivate;
 
-    /// \addtogroup gazebo_gui
+    /// \addtogroup gazebo_gui GUI
+    /// \brief The graphical user interface
     /// \{
 
     /// \class CloneWindow CloneWindow.hh gui/CloneWindow.hh