+++ gazebo/gui/model/LinkInspector.cc
@@ -175,6 +175,8 @@ void LinkInspector::SetLinkId(const std::string &_id)
 /////////////////////////////////////////////////
 void LinkInspector::Open()
 {
+  emit this->Opened();
+
   this->dataPtr->linkConfig->Init();
   this->dataPtr->visualConfig->Init();
   this->dataPtr->collisionConfig->Init();